requirements.txt
firebase-admin
fastapi
uvicorn
psycopg2-binary
sqlalchemy
alembic
python-dotenv
